Addressing comments from MR, adding some tests for DDCAD files
