--- FCCee/compact/FCCee_IDEA_o1_v01/FCCee_IDEA_o1_v01.xml
@@ -1,49 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<lccdd xmlns:compact="http://www.lcsim.org/schemas/compact/1.0"
-  xmlns:xs="http://www.w3.org/2001/XMLSchema"
-  xs:noNamespaceSchemaLocation="http://www.lcsim.org/schemas/compact/1.0/compact.xsd">
-
-
-
-  <info name="FCCDectMaster"
-    title="FCCee IDEA detector geometry master file"
-    author="Niloufar Alipour Tehrani"
-    url="no"
-    status="development"
-    version="1.0">
-    <comment> 
-      One stop shop for a baseline / the latest developments of the FCCee Detector concept.
-    </comment>
-  </info>
-
-  <include ref="${DD4hepINSTALL}/DDDetectors/compact/detector_types.xml" />
-
-  <includes>
-    <gdmlFile  ref="elements_o1_v01.xml"/>
-    <gdmlFile  ref="materials_o1_v01.xml"/>
-  </includes>
-
-  <define>
-    <constant name="world_size" value="25*m"/>
-    <constant name="world_x" value="world_size"/>
-    <constant name="world_y" value="world_size"/>
-    <constant name="world_z" value="world_size"/>
-  </define>
-
-  <!-- Import dimensions -->
-  <include ref="DectDimensions_IDEA_o1_v01.xml"/>
-
-  <!-- Import accelerator components -->
-  <include ref="Beampipe_o1_v01.xml"/>
-  <include ref="BeamInstrumentation_o1_v01.xml"/>
-  <include ref="HOMAbsorber_o1_v01.xml"/>
-  <include ref="LumiCal_o1_v01.xml"/>
-
-  <!-- Import vertex -->
-  <include ref="Vertex_IDEA_o1_v01.xml"/>
-
-  <plugins>
-    <plugin name="DD4hepVolumeManager"/>
-    <plugin name="InstallSurfaceManager"/>
-  </plugins>
-</lccdd>
