--- FCCee/compact/FCCee_IDEA_o1_v01/scripts/check_material_budget.sh
@@ -1,2 +1,0 @@
-k4run scripts/utils/material_scan.py 
-python scripts/utils/material_plots.py --fname out_material_scan.root --etaMax 3 --x0max 0.18 # eta of 3 is equal to theta of 0.099, so 99 mrad
