--- FCCee/compact/FCCee_IDEA_o1_v01/scripts/teveDisplay.sh
@@ -1,8 +1,0 @@
-### Script to look at detector geometry (warning: very slow for complex detectors)
-### and over bad connection
-
-# IDEA Vertex
-# teveDisplay -ui -compact FCCee_IDEA_o1_v01.xml 
-
-# IDEA Vertex IDEA
-teveDisplay -ui -compact FCCee_IDEA_Empty_o1_v01.xml Vertex_IDEA_o1_v01.xml
