--- FCCee/compact/FCCee_IDEA_o1_v01/scripts/utils/overlap.mac
@@ -1,2 +1,0 @@
-/geometry/test/run
-exit
