+++ FCCee/compact/IDEA/IDEA_o1_v01/FCCee_IDEA_Empty_o1_v01.xml
@@ -1,0 +1,28 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<lccdd xmlns:compact="http://www.lcsim.org/schemas/compact/1.0"
+  xmlns:xs="http://www.w3.org/2001/XMLSchema"
+  xs:noNamespaceSchemaLocation="http://www.lcsim.org/schemas/compact/1.0/compact.xsd">
+
+  <includes>
+    <gdmlFile  ref="elements_o1_v01.xml"/>
+    <gdmlFile  ref="materials_o1_v01.xml"/>
+  </includes>
+
+  <info name="FCCDectMaster"
+    title="master file with includes and world dimension"
+    author="C. Helsens"
+    url="no"
+    status="development"
+    version="1.0">
+    <comment>Use this one if you want to use official dimensions but only place one detector inside</comment>
+  </info>
+  <define>
+    <constant name="world_size" value="8000*mm"/>
+    <constant name="world_x" value="world_size"/>
+    <constant name="world_y" value="world_size"/>
+    <constant name="world_z" value="world_size"/>
+  </define>
+
+  <include ref="DectDimensions_IDEA_o1_v01.xml" />
+
+</lccdd>
