+++ FCCee/compact/IDEA/IDEA_o1_v01/scripts/utils/overlap.mac
@@ -1,0 +1,2 @@
+/geometry/test/run
+exit
